+++ custom_components/volcano_hybrid/volcano_ble/const.py
@@ -51,3 +51,4 @@ class VolcanoSensor(StrEnum):
     PRJ5 = "prj5"
     HIST1 = "hist1"
     HIST2 = "hist2"
+    LAST_FAULT = "last_fault"